Fix null-dereference and narrowing-conversion findings from clang-tidy
The SQLite persistence-restore loops in btech_persistence_sqlite.c
track a "current group" object pointer (map/mech/turret/autopilot)
keyed off a DbRef, initialized via current_X = NOTHING. A malformed
row whose grouping dbref equals NOTHING would bypass the
lookup-and-null-check block, leaving the pointer NULL when
dereferenced later in the same loop iteration. clang-analyzer's
core.NullDereference check flagged 16 of these; 2 more instances of
the identical pattern (mech_criticals, map_objects) weren't flagged
but share the same root cause. All 18 are now rejected at
row-validation time by checking the grouping dbref against NOTHING,
consistent with the file's existing "!= NOTHING && !is_good_obj(...)"
validation idiom for optional reference columns.

Also fixes a bugprone-narrowing-conversions warning in mkindx.c by
explicitly casting strlen()'s size_t result to int (the buffer is a
fixed 400-byte line, so no truncation risk).

+++ src/mkindx/mkindx.c
@@ -33,7 +33,7 @@ int main(int argc, char *argv[]) {
   while (fgets(line, LINE_SIZE, rfp) != NULL) {
     ++lineno;
 
-    n = strlen(line);
+    n = (int)strlen(line);
     if (line[n - 1] != '\n') {
       fprintf(stderr, "line %d: line too long\n", lineno);
     }
